requirements.txt
# requirements.txt - List of Python dependencies for the Signal Bot project.
# This file contains the external packages required for running the bot and its tests.
python-dotenv>=0.20.0
pytest>=7.0.0
pytest-asyncio>=0.18.0

# End of requirements.txt